pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-friendly-net-detection-server"
version = "0.1.0"
description = "Friendly Network Detection server"
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Apfelwurm", email = "Alexander@volzit.de"}]
license = {text = "MIT"}
dependencies = [
  "PyYAML",
  "cryptography"
]

[project.scripts]
open-friendly-net-detection-server = "open_friendly_net_detection_server.server:main"
